[CI][Nightly] tell a stuck benchmark file apart from an expensive one (#1864)
Closes #1813.

## Why

The nightly benchmark job ([run
31272374766](https://github.com/tile-ai/TileOPs/actions/runs/31272374766))
was cancelled at its 150-minute limit having run 34 of 59 files. Because
GitHub cancelled it, `bench_results.xml` was never written and
`publish-bench-data` skipped — 149 minutes of exclusive GPU time
produced no data at all.

`run_benchmarks.py` had one knob, `--timeout-per-file`, doing two jobs:
it bounds a native failure, and it also kills a file for being
expensive. Three files reached the 900s cap while running normally.
Given the full time they need, none is anywhere near a hang:

| File | Full cold-cache run | Slowest single test | Under the cap |
|---|---|---|---|
| `bench_convolution.py` | 3838s | 100s | killed, 11 of 57 tests |
| `bench_gqa.py` | 1153s | 76s | killed, 27 of 34 tests |
| `bench_gated_deltanet_prefill.py` | 1129s | 98s | killed, 20 of 26
tests |

The cap was also self-perpetuating. `bench_convolution.py` spends 92% of
its runtime autotuning (3838s tuned, 324s untuned), and tilelang caches
autotune results across processes — the same tuned conv2d subset takes
**1035s cold and 8s warm**. A file killed mid-tune writes nothing, so it
arrived cold every night and hit the cap again. The benchmarks do not
need to get cheaper; they need to be allowed to finish once.

## The change

Two limits with separate jobs replace the one.

- **`--stall-timeout`** measures time since the child last *started a
test* — a pytest plugin rewrites a heartbeat file, the parent stats its
mtime in the poll loop it already runs. A hang is caught as fast as
before, and the kill message names the test.
- **`--total-budget`** stops launching files, writes the merged report,
and records what it never reached as skipped, instead of letting CI
cancel the job and discard everything.

Sized from a measured full sweep: 59 files, cold cache, idle H200,
**18438s (5h07m)**; median file 86s, p90 673s. `--total-budget 21600`
covers it with margin and `timeout-minutes 380` stays above the budget.

## Defects the same investigation exposed

**Failure output was being lost.** The child wrote its exit code to the
status pipe before flushing stdout. Redirected to a file, stdout is
block-buffered, so the parent printed the log while pytest's `FAILURES`
section was still in the buffer — nightly reported failures with no
reason attached to any of them.

**20 benchmark failures, none of which had a visible cause.** Four bench
files (`bench_binary_elementwise`, `bench_fused_moe_experts`,
`bench_moe_fused_moe`, `bench_moe_fused_activation`) passed `dtype=` to
constructors that bind dtype per call. The last two are in the region
the cancelled nightly never reached, and only turned up in the full
sweep.

**One op bug.** `_ArgreduceOpBase._prepare_input` skipped the shared
input validation on its strided fast path, so the op returned a correct
result and then failed `eval_roofline()` for an unbound dtype, with
device and rank unchecked.

## Testing

Verified in `ghcr.io/tile-ai/tileops-runner:afcebed1-torch2.10-dev`, the
CI-identical image:

- `benchmarks/tests/test_run_benchmarks.py` — 8 passed, covering the
flag rename plus the three cases the split introduces
- the five formerly-failing bench files plus
`tests/ops/test_argreduce.py` — 196 passed, was 20 failed (two runs: 188
before the sweep surfaced the last two files, 8 after)
- full 59-file sweep under the new runner — no stalls, no crashes, every
file completed

+++ benchmarks/ops/bench_moe_fused_moe.py
@@ -119,7 +119,7 @@ def _run_bench(
         num_tokens=num_tokens, num_experts=num_experts, top_k=top_k,
         hidden_size=hidden_size, ffn_size=ffn_size,
         scoring_func=scoring_func, renormalize=renormalize,
-        routed_scaling_factor=routed_scaling_factor, dtype=dtype,
+        routed_scaling_factor=routed_scaling_factor,
     )
 
     if with_correction_bias:
